removed stream setup and refactor
